gh-229  fixing confirmation counting
